+++ SharpSource/SharpSource/Utilities/DiagnosticId.cs
@@ -67,4 +67,5 @@ public static class DiagnosticId
     public const string LoggerMessageAttribute = "SS065";
     public const string DisposableFieldIsNotDisposed = "SS066";
     public const string RedisResponseNotHandled = "SS067";
+    public const string TimeSpanConstructedWithTicks = "SS068";
 }